requirements (1).txt
streamlit==1.48.0
pandas==2.3.1
numpy==2.3.2
python-docx==1.1.2
